requirements.txt
opencv-python
ultralytics
torch
torchvision
transformers
pillow
pyttsx3
numpy
pytesseract
speechrecognition